解决bug

+++ input/read_data.py
@@ -5,7 +5,7 @@ import pandas as pd
 def read_data(train_path, test_path):
     dtypes_train = {
         'TERMINALNO': 'uint32',
-        'TIME': 'uint16',
+        'TIME': 'uint64',
         'TRIP_ID': 'uint8',
         'LATITUDE': 'float32',
         'LONGITUDE': 'float32',
@@ -17,7 +17,7 @@ def read_data(train_path, test_path):
     }
     dtypes_test = {
         'TERMINALNO': 'uint32',
-        'TIME': 'uint16',
+        'TIME': 'uint64',
         'TRIP_ID': 'uint8',
         'LATITUDE': 'float32',
         'LONGITUDE': 'float32',
@@ -28,4 +28,5 @@ def read_data(train_path, test_path):
     }
     train = pd.read_csv(train_path, encoding='utf8', dtype=dtypes_train)
     test = pd.read_csv(test_path, encoding='utf8', dtype=dtypes_test)
+
     return train, test
